🌱 Split SetupInstructionsDialog into co-located sub-components
Extract the constants/OAuth step table, the copy-state hook, the shared
CopyableCommand/CollapsibleGuide primitives, and the architecture,
dev-mode, Kubernetes, security and OAuth sections into sibling modules.
SetupInstructionsDialog.tsx goes from 411 to 143 lines / 6 hooks with no
behaviour change.
